refactor: TimerSession ID 타입을 Int에서 Long으로 변경

+++ core/domain/src/main/kotlin/com/jm/focustimer/domain/usecase/session/ManageTimerSessionUseCase.kt
@@ -36,7 +36,7 @@ class ManageTimerSessionUseCase @Inject constructor(
      * 세션을 완료 상태로 업데이트합니다
      */
     suspend fun completeSession(
-        sessionId: Int,
+        sessionId: Long,
         presetId: Int? = null,
         startTime: Instant?,
         initialDuration: Duration,
@@ -63,7 +63,7 @@ class ManageTimerSessionUseCase @Inject constructor(
      * 세션을 미완료 상태로 종료합니다 (중도 정지)
      */
     suspend fun stopSession(
-        sessionId: Int,
+        sessionId: Long,
         presetId: Int? = null,
         startTime: Instant?,
         initialDuration: Duration,
